Move when the border color is set
Originally, the border color was set on `setup()` which means during the
initialization of the view.

Then, when changing the interface style, the color value remained the same,
given how semantic colors resolve.

For that reason, I moved setting the color to the `drawBorder` method that
in turns is called in `layoutSubviews`, where the trait colleciton is correctly set
then the color will resolve correctly.

More information: https://stackoverflow.com/a/57177411/611678